requirements.txt
torch==1.9.1 # pip install torch --extra-index-url https://download.pytorch.org/whl/cu113
#pip install torch==1.11.0+cu113 torchvision==0.12.0+cu113 torchaudio==0.11.0 --extra-index-url https://download.pytorch.org/whl/cu113
numpy 
sklearn
pytorch-lightning==1.6.0
pytorch_msssim==0.2.1
monai==0.8.1
torchmetrics==0.11.4
torch-fidelity
torchio==0.18.70
pillow==9.0.0
einops==0.4.1
torchvision
matplotlib
pandas 
lpips

streamlit